+++ packages/palette/src/elements/Cards/CardTag.shared.tsx
@@ -1,6 +1,6 @@
 import { StyleProp, ViewStyle } from "react-native"
 
-export interface TagProps {
+export interface CardTagProps {
   text: string
   textColor: string
   color: string
