+++ packages/palette/src/elements/Cards/LargeCard.shared.tsx
@@ -1,8 +1,8 @@
-import { TagProps } from "./Tag.shared"
+import { CardTagProps } from "./CardTag.shared"
 
 export interface LargeCardProps {
   image: string
   title: string
   subtitle?: string
-  tag?: TagProps
+  tag?: CardTagProps
 }
