+++ packages/palette/src/elements/Cards/MediumCard.shared.tsx
@@ -1,8 +1,8 @@
-import { TagProps } from "./Tag.shared"
+import { CardTagProps } from "./CardTag.shared"
 
 export interface MediumCardProps {
   image: string
   title: string
   subtitle?: string
-  tag?: TagProps
+  tag?: CardTagProps
 }
