+++ packages/palette/src/elements/Cards/SmallCard.shared.tsx
@@ -1,8 +1,8 @@
-import { TagProps } from "./Tag.shared"
+import { CardTagProps } from "./CardTag.shared"
 
 export interface SmallCardProps {
   images: string[]
   title: string
   subtitle?: string
-  tag?: TagProps
+  tag?: CardTagProps
 }
